Add files via upload
Developed a calculator with the basic functionalities i.e., add, subtract, multiply and divide. 

Each of the functionality is made as an individual class in a separate package. That is,
 Addition functionality is in a package: calculator.addition
 Subtraction functionality is in a package: calculator.subtract
 Multiplication functionality is in a package: calculator.multiply
 Division functionality is in a package: calculator.division
A Driver class is in a separate package, imported all the packages in it, and shown the execution of the program.